requirements.txt
langchain
langchain-groq
langchain-community
sentence-transformers
chromadb
pypdf
python-dotenv
tqdm
streamlit
